audio/requirements.txt
Flask
numpy
librosa
opencc
pyannote.audio
faster-whisper
werkzeug
ffmpeg-python
torch
torchaudio
